Fix auth and sync configuration - Refactor to remove config service dependency - Attachement resolver configuration - Refactor + clean up extract resouce mapper validation
